Move utility code to the library package
